// src/types.ts
export type TabType = 
  | 'beranda'
  | 'semua-turnamen'
  | 'ff'
  | 'mlbb'
  | 'total-hadiah'
  | 'pengumuman'
  | 'riwayat'
  | 'status-pembayaran'
  | 'form-pendaftaran'
  | 'pengaturan-umum'
  | 'bantuan'
  | 'arsip'
  | 'bagikan'
  | 'cara-daftar'
  | 'aturan'
  | 'tim'
  | 'info-match'
  | 'grup'
  | 'laporan'
  | 'kontak'
  | 'login'
  | 'prediksi'
  | 'saldo'
  | 'admin'
  | 'topup-game'
  | 'unduh-apk'
  | 'donasi'
  | 'gemini-ai';
// ...
export interface AnnouncementItem {
  id: string;
  title: string;
  content: string;
  date: string;
  category: 'Info Penting' | 'Perubahan Jadwal' | 'Pengingat' | 'Umum' | string;
  isImportant?: boolean;
  targetAudience?: 'Semua Kapten & Member' | 'Kapten Free Fire' | 'Kapten MLBB' | 'Kapten Spesifik' | string;
  targetTeamId?: string;
  targetTeamName?: string;
  targetCaptainName?: string;
  targetCaptainPhone?: string;
  targetGame?: 'FF' | 'MLBB' | 'Semua' | string;
  deviceSentCount?: number;
}